pyproject.toml
[project]
name = "agentic-research-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.96.0",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langgraph>=1.1.8",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
]
